requirements.txt
fastapi==0.104.1
regex==2023.10.3
mysql-connector-python==8.2.0
requests==2.31.0
uvicorn==0.24.0.post1